optical-character-recognition-api/requirements.txt
Flask==2.1.2
gunicorn==20.1.0
opencv-python-headless==4.5.5.64
pytesseract==0.3.9
Pillow==9.1.1
google-cloud-storage==2.3.0
numpy==1.22.4
tensorflow==2.9.1
gdown==4.4.0
nltk==3.7